package.json: Important modifications.
We want comments in .d.ts files, of course, but not in .js files.
Unfortunately, tsc's --removeComments removes comments
from any output (how stupid!), so we generate .js and .d.ts files
separately:
  - Added scripts "generateJS" and "generateDts".
  - Changed script "compile" to first remove all files within outDir,
    then run scripts "generateJs" and "generateDts".
	The removal of the files is necessary because tsc refuses to
	overwrite existing .d.ts files in outDir because it considers them
	to belong to the input, for whatever reason.

Added "config" object with properties "rootDir" and "outDir"
to be passed to tsc via command line options.
This allows us to use the same tsconfig.json for multiple projects
without using inheritance.
Furthermore, it is convenient to have the property defining outDir
and those needing that same dir to be specified ("main" and "types")
in the same file.

devDependencies:
  - Removed @microsoft/api-extractor (not the right tool for my purpose).
  - Added typedoc and typedoc-plugin-markdown.

Added script "generateDoc" to invoke typedoc
and changed script "compile" to run it at the end.